Add Cast From JSON::Type Support
While testing the Amber master branch we found that there is an error
when passing a `JSON::Type` to a model.

This corrects by adding a `cast_json_to_field` method that take in a
`JSON::Type` and asserts it to the correct type.

Backwards maintainbility when using a Hash of string values or DB::Any.